requirements.txt
absl-py==2.1.0
aiosignal==1.3.1
alabaster==0.7.16
annotated-types==0.6.0
asttokens==2.4.1
attrs==23.2.0
autopep8==2.1.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
category-encoders==2.6.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
ConfigSpace==0.7.1
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
dgl==2.1.0
dglgo==0.0.2
docopt==0.6.2
docutils==0.20.1
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.9.0
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2023.4.0
gensim==4.3.2
grpcio==1.62.2
hpbandster==0.7.4
idna==3.7
imagesize==1.4.1
ipython==8.12.3
isort==5.13.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.4.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
littleutils==0.2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
netifaces==0.11.0
networkx==3.2.1
numpy==1.26.3
numpydoc==1.7.0
ogb==1.3.6
outdated==0.2.2
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
patsy==0.5.6
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==4.2.1
prompt-toolkit==3.0.43
protobuf==5.26.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.2
pycodestyle==2.11.1
pydantic==2.7.0
pydantic_core==2.18.1
Pygments==2.17.2
pyparsing==3.1.2
Pyro4==4.82
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.2
ray==2.11.0
rdkit-pypi==2022.9.5
referencing==0.34.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scikit-learn==1.4.2
scipy==1.13.0
serpent==1.41
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
snowballstemmer==2.2.0
soupsieve==2.5
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
stack-data==0.6.3
statsmodels==0.14.1
sympy==1.12
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
threadpoolctl==3.4.0
tinycss2==1.3.0
torch==2.2.1+cpu
torchaudio==2.2.1+cpu
torchdata==0.7.1
torchvision==0.17.1+cpu
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
typer==0.12.3
typing_extensions==4.8.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.2
wrapt==1.16.0
yarg==0.1.9
